requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Machine Learning
lightgbm>=3.3.0
shap>=0.41.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Model Persistence
joblib>=1.1.0

# AI/LLM Integration
openai-agents>=0.0.9

# Environment Management
python-dotenv>=0.19.0

# Optional: Development and Testing
pytest>=7.0.0
jupyter>=1.0.0
notebook>=6.0.0

# Optional: Enhanced Analytics
plotly>=5.0.0
scipy>=1.8.0